refactor(core): extract the protected-path classifiers from _hooklib (#321, slice 2/4)
Second of four. _hooklib goes 1,050 -> 880 lines; it was 1,263 before the
partition began, so two slices have taken 30% off the god module.

_protectedlib owns one question - "what class of protected thing is this path?" -
across append-only audit logs (H-05), ADR decision files (H-11), the activation
manifest, and the gate-marker directory.

They belong together because the guards branch on the answer as a SET.
classify_protected returns EVERY class a path hits, which is load-bearing:
#528/#529 showed that adding decision-log.md to the audit set while leaving it in
the decisions set STILL blocked, because pre-write.py checks the classes
independently. Splitting these across modules would let one classification move
without the others and reopen exactly that.

Same measurement as slice 1: the cluster referenced exactly ONE symbol from the
rest of _hooklib, and NOTHING in the rest referenced the cluster.

That one symbol was repo_rel, which moved to the _pathnorm floor rather than
being imported back (which would have been a cycle). It earns its place there on
the same test norm_path did: it references no module symbol at all, and the
remaining slices need it too. Its docstring is also the canonical statement of
why a LEXICAL relpath is wrong here - the 8.3-short-name and symlink divergence
that independently produced #539 and #541 this week.

_hooklib re-exports every moved name, so consumers are UNCHANGED and parity rests
on 1,185 pre-existing hook tests that did not move. Public surface verified
against origin/main: nothing lost, nothing gained.

No crypto moved. H-09b nevertheless blocked this commit, and the cause turned out
to be worth more than the slice: the .git hook was running a ca-codex 0.3.0
cached enforcer that PREDATES #279, so it scanned the audit log and matched the
detector's own "Crypto/TLS pattern detected" message text - the exact
self-perpetuating false positive #279 closed. Filed as #556; the shim is
repointed at this checkout's enforcer, which carries the exemption. That
STRENGTHENS the backstop rather than bypassing it - no override was used.

`import subprocess` in _hooklib now looks dead to a naive scan but already
carries a noqa explaining it is kept so an external consumer referencing
_hooklib.subprocess keeps resolving. Pre-existing, not collateral.

ca 2.10.5 -> 2.10.6, ca-codex 0.3.4 -> 0.3.5, ca-pi 0.1.40 -> 0.1.41.
